+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Improvements
 - Fallback when server notification streams are turned off ([#967](../../pull/967))
+- Show and edit yaml and json files using codemirror ([#969](../../pull/969))
 
 ## 1.17.0